JavaWeb/property-jhipster/test_room/src/main/webapp/app/entities/name/update/name-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { NameService } from '../service/name.service';
import { IName, Name } from '../name.model';

import { NameUpdateComponent } from './name-update.component';

describe('Name Management Update Component', () => {
  let comp: NameUpdateComponent;
  let fixture: ComponentFixture<NameUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let nameService: NameService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [NameUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(NameUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(NameUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    nameService = TestBed.inject(NameService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should update editForm', () => {
      const name: IName = { id: 456 };

      activatedRoute.data = of({ name });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(name));
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Name>>();
      const name = { id: 123 };
      jest.spyOn(nameService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ name });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: name }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(nameService.update).toHaveBeenCalledWith(name);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Name>>();
      const name = new Name();
      jest.spyOn(nameService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ name });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: name }));
      saveSubject.complete();

      // THEN
      expect(nameService.create).toHaveBeenCalledWith(name);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Name>>();
      const name = { id: 123 };
      jest.spyOn(nameService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ name });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(nameService.update).toHaveBeenCalledWith(name);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });
});
